Split out Request, deprecate methods
